requirements.txt
# Core Dependencies
langchain==0.1.0
langchain-community==0.0.13
pypdf2==3.0.1
pymupdf==1.23.8
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3

# LLM APIs and Models
openai==1.6.1
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0
bitsandbytes==0.41.3

# Evaluation Metrics
evaluate==0.4.1
bert-score==0.3.13
rouge-score==0.1.2
nltk==3.8.1
sacrebleu==2.3.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
scikit-learn==1.3.2
